+++ spec/localeapp/configuration_spec.rb
@@ -81,7 +81,7 @@ describe Localeapp::Configuration do
 
   context "disabled_sending_environments" do
     it "sets deprecated_environment_config_used flag to true" do
-      configuration.disabled_sending_environments = %(foo bar)
+      configuration.disabled_sending_environments = %w(foo bar)
       configuration.deprecated_environment_config_used?.should be_true
     end
 
@@ -108,7 +108,7 @@ describe Localeapp::Configuration do
 
   context "disabled_reloading_environments" do
     it "sets deprecated_environment_config_used flag to true" do
-      configuration.disabled_reloading_environments = %(foo bar)
+      configuration.disabled_reloading_environments = %w(foo bar)
       configuration.deprecated_environment_config_used?.should be_true
     end
 
@@ -135,7 +135,7 @@ describe Localeapp::Configuration do
   
   context "disabled_polling_environments" do
     it "sets deprecated_environment_config_used flag to true" do
-      configuration.disabled_polling_environments = %(foo bar)
+      configuration.disabled_polling_environments = %w(foo bar)
       configuration.deprecated_environment_config_used?.should be_true
     end
 
@@ -163,27 +163,39 @@ describe Localeapp::Configuration do
   describe "polling_disabled?" do
     context "deprecated syntax used" do
       it "is true when environment is disabled" do
-        configuration.disabled_polling_environments = %(foo)
+        configuration.disabled_polling_environments = %w(foo)
         configuration.environment_name = 'foo'
         configuration.should be_polling_disabled
       end
 
       it "is false when environment is not disabled" do
-        configuration.disabled_polling_environments = %(foo)
+        configuration.disabled_polling_environments = %w(foo)
         configuration.environment_name = 'bar'
         configuration.should_not be_polling_disabled
+      end
+
+      it "supports symbols in list of environments" do
+        configuration.disabled_polling_environments = [:foo]
+        configuration.environment_name = 'foo'
+        configuration.should be_polling_disabled
       end
     end
 
     context "new syntax used" do
       it "is true when environment is not enabled" do
-        configuration.polling_environments = %(foo)
+        configuration.polling_environments = %w(foo)
         configuration.environment_name = 'bar'
         configuration.should be_polling_disabled
       end
 
       it "is false when environment is enabled" do
-        configuration.polling_environments = %(foo)
+        configuration.polling_environments = %w(foo)
+        configuration.environment_name = 'foo'
+        configuration.should_not be_polling_disabled
+      end
+
+      it "supports symbols in list of environments" do
+        configuration.polling_environments = [:foo]
         configuration.environment_name = 'foo'
         configuration.should_not be_polling_disabled
       end
@@ -193,27 +205,39 @@ describe Localeapp::Configuration do
   describe "reloading_disabled?" do
     context "deprecated syntax used" do
       it "is true when environment is disabled" do
-        configuration.disabled_reloading_environments = %(foo)
+        configuration.disabled_reloading_environments = %w(foo)
         configuration.environment_name = 'foo'
         configuration.should be_reloading_disabled
       end
 
       it "is false when environment is not disabled" do
-        configuration.disabled_reloading_environments = %(foo)
+        configuration.disabled_reloading_environments = %w(foo)
         configuration.environment_name = 'bar'
         configuration.should_not be_reloading_disabled
+      end
+
+      it "supports symbols in list of environments" do
+        configuration.disabled_reloading_environments = [:foo]
+        configuration.environment_name = 'foo'
+        configuration.should be_reloading_disabled
       end
     end
 
     context "new syntax used" do
       it "is true when environment is not enabled" do
-        configuration.reloading_environments = %(foo)
+        configuration.reloading_environments = %w(foo)
         configuration.environment_name = 'bar'
         configuration.should be_reloading_disabled
       end
 
       it "is false when environment is enabled" do
-        configuration.reloading_environments = %(foo)
+        configuration.reloading_environments = %w(foo)
+        configuration.environment_name = 'foo'
+        configuration.should_not be_reloading_disabled
+      end
+
+      it "supports symbols in list of environments" do
+        configuration.reloading_environments = [:foo]
         configuration.environment_name = 'foo'
         configuration.should_not be_reloading_disabled
       end
@@ -223,27 +247,39 @@ describe Localeapp::Configuration do
   describe "sending_disabled?" do
     context "deprecated syntax used" do
       it "is true when environment is disabled" do
-        configuration.disabled_sending_environments = %(foo)
+        configuration.disabled_sending_environments = %w(foo)
         configuration.environment_name = 'foo'
         configuration.should be_sending_disabled
       end
 
       it "is false when environment is not disabled" do
-        configuration.disabled_sending_environments = %(foo)
+        configuration.disabled_sending_environments = %w(foo)
         configuration.environment_name = 'bar'
         configuration.should_not be_sending_disabled
+      end
+
+      it "supports symbols in the list of enviroments" do
+        configuration.disabled_sending_environments = [:foo]
+        configuration.environment_name = 'foo'
+        configuration.should be_sending_disabled
       end
     end
 
     context "new syntax used" do
       it "is true when environment is not enabled" do
-        configuration.sending_environments = %(foo)
+        configuration.sending_environments = %w(foo)
         configuration.environment_name = 'bar'
         configuration.should be_sending_disabled
       end
 
       it "is false when environment is enabled" do
-        configuration.sending_environments = %(foo)
+        configuration.sending_environments = %w(foo)
+        configuration.environment_name = 'foo'
+        configuration.should_not be_sending_disabled
+      end
+
+      it "supports symbols in the list of environments" do
+        configuration.sending_environments = [:foo]
         configuration.environment_name = 'foo'
         configuration.should_not be_sending_disabled
       end